LayoutTests/media/modern-media-controls/compact-media-controls/compact-media-controls-layout.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>CompactMediaControls</code> layout.");

const mediaControls = new CompactMediaControls;

debug("With controls at their default size only the play/pause button is showing with a center style.");
shouldBe("mediaControls.children", "[mediaControls.playPauseButton]");
shouldBe("mediaControls.playPauseButton.style", "Button.Styles.Center");
debug("");

debug("With controls just under the minimum size to show prominent controls only the play/pause button is showing with a small-center style.");
mediaControls.width = MaximumSizeToShowSmallProminentControl - 1;
mediaControls.height = MaximumSizeToShowSmallProminentControl - 1;
shouldBe("mediaControls.children", "[mediaControls.playPauseButton]");
shouldBe("mediaControls.playPauseButton.style", "Button.Styles.SmallCenter");
debug("");

debug("With controls just under the minimum size to show any controls there are no children.");
mediaControls.width = MinimumSizeToShowAnyControl - 1;
mediaControls.height = MinimumSizeToShowAnyControl - 1;
shouldBe("mediaControls.children.length", "0");
debug("");

debug("With controls above the minimum size and the error placard set only the invalid placard is shown.");
mediaControls.width = MaximumSizeToShowSmallProminentControl;
mediaControls.height = MaximumSizeToShowSmallProminentControl;
mediaControls.placard = mediaControls.invalidPlacard;
shouldBe("mediaControls.children", "[mediaControls.invalidPlacard]");
shouldBe("mediaControls.invalidPlacard.width", "mediaControls.width");
shouldBe("mediaControls.invalidPlacard.height", "mediaControls.height");
debug("");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
